Slice 8 — Tests

+++ ProductCatalog.Infrastructure/Cache/CacheSettings.cs
@@ -2,5 +2,5 @@ namespace ProductCatalog.Infrastructure.Cache;
 
 public class CacheSettings
 {
-    public int ProductTtlMinutes { get; set; } = 5;
+    public double ProductTtlMinutes { get; set; } = 5;
 }
